+++ src/components/SurveyModal.tsx
@@ -1,12 +1,14 @@
 'use client';
 
 import { useState, useEffect } from 'react';
-import { useForm } from 'react-hook-form';
+import { useForm, FormProvider } from 'react-hook-form';
 import { motion, AnimatePresence } from 'framer-motion';
 import { X, ChevronLeft, ChevronRight, Send } from 'lucide-react';
 import { Button } from '@/components/ui/button';
 import { SurveyQuestion, PackageSurvey, sportEquipmentOptions } from '@/data/surveys';
 import { cn } from '@/lib/utils';
+import SurveyQuestionRenderer from './SurveyQuestionRenderer';
+import SurveySuccessPage from './SurveySuccessPage';
 
 interface SurveyModalProps {
   survey: PackageSurvey;
@@ -21,7 +23,8 @@ export default function SurveyModal({ survey, isOpen, onClose, onSubmit }: Surve
   const [showSuccessPage, setShowSuccessPage] = useState(false);
   const [submittedData, setSubmittedData] = useState<Record<string, unknown> | null>(null);
 
-  const { register, handleSubmit, watch, formState: { errors }, reset } = useForm();
+  const methods = useForm();
+  const { register, handleSubmit, watch, formState: { errors }, reset } = methods;
 
   // Spor türü değişikliklerini izle
   const watchedSportType = watch('sport-type');
@@ -104,242 +107,6 @@ export default function SurveyModal({ survey, isOpen, onClose, onSubmit }: Surve
     setSubmittedData(null);
   };
 
-  const renderQuestion = (question: SurveyQuestion) => {
-    const watchedValue = watch(question.id);
-
-    switch (question.type) {
-      case 'text':
-        return (
-          <div className="space-y-2">
-            <input
-              {...register(question.id, { required: question.required })}
-              type="text"
-              placeholder={question.placeholder}
-              className="w-full px-4 py-3 border border-gray-300 rounded-lg focus:ring-2 focus:ring-blue-500 focus:border-transparent transition-all"
-            />
-            {errors[question.id] && (
-              <p className="text-red-500 text-sm">Bu alan zorunludur</p>
-            )}
-          </div>
-        );
-
-      case 'textarea':
-        return (
-          <div className="space-y-2">
-            <textarea
-              {...register(question.id, { required: question.required })}
-              placeholder={question.placeholder}
-              rows={4}
-              className="w-full px-4 py-3 border border-gray-300 rounded-lg focus:ring-2 focus:ring-blue-500 focus:border-transparent transition-all resize-none"
-            />
-            {errors[question.id] && (
-              <p className="text-red-500 text-sm">Bu alan zorunludur</p>
-            )}
-          </div>
-        );
-
-      case 'radio':
-        return (
-          <div className="space-y-3">
-            {question.options?.map((option, index) => (
-              <label
-                key={index}
-                className={cn(
-                  "flex items-center p-4 border rounded-lg cursor-pointer transition-all hover:bg-gray-50",
-                  watchedValue === option ? "border-blue-500 bg-blue-50" : "border-gray-200"
-                )}
-              >
-                <input
-                  {...register(question.id, { required: question.required })}
-                  type="radio"
-                  value={option}
-                  className="sr-only"
-                />
-                <div className={cn(
-                  "w-4 h-4 rounded-full border-2 mr-3 flex items-center justify-center",
-                  watchedValue === option ? "border-blue-500" : "border-gray-300"
-                )}>
-                  {watchedValue === option && (
-                    <div className="w-2 h-2 rounded-full bg-blue-500" />
-                  )}
-                </div>
-                <span className="text-gray-700">{option}</span>
-              </label>
-            ))}
-            {errors[question.id] && (
-              <p className="text-red-500 text-sm">Bu alan zorunludur</p>
-            )}
-          </div>
-        );
-
-      case 'multiple-choice':
-        return (
-          <div className="space-y-3">
-            {question.options?.map((option, index) => (
-              <label
-                key={index}
-                className={cn(
-                  "flex items-center p-4 border rounded-lg cursor-pointer transition-all hover:bg-gray-50",
-                  watchedValue?.includes(option) ? "border-blue-500 bg-blue-50" : "border-gray-200"
-                )}
-              >
-                <input
-                  {...register(question.id)}
-                  type="checkbox"
-                  value={option}
-                  className="sr-only"
-                />
-                <div className={cn(
-                  "w-4 h-4 rounded border-2 mr-3 flex items-center justify-center",
-                  watchedValue?.includes(option) ? "border-blue-500 bg-blue-500" : "border-gray-300"
-                )}>
-                  {watchedValue?.includes(option) && (
-                    <svg className="w-3 h-3 text-white" fill="currentColor" viewBox="0 0 20 20">
-                      <path fillRule="evenodd" d="M16.707 5.293a1 1 0 010 1.414l-8 8a1 1 0 01-1.414 0l-4-4a1 1 0 011.414-1.414L8 12.586l7.293-7.293a1 1 0 011.414 0z" clipRule="evenodd" />
-                    </svg>
-                  )}
-                </div>
-                <span className="text-gray-700">{option}</span>
-              </label>
-            ))}
-            {errors[question.id] && (
-              <p className="text-red-500 text-sm">Bu alan zorunludur</p>
-            )}
-          </div>
-        );
-
-      default:
-        return null;
-    }
-  };
-
-  // Anket sonuç sayfası render fonksiyonu
-  const renderSuccessPage = () => {
-    const getSportEquipment = () => {
-      if (submittedData?.['sport-equipment'] && Array.isArray(submittedData['sport-equipment'])) {
-        return submittedData['sport-equipment'].filter((item: string) => item !== 'Hiçbirini istemiyorum');
-      }
-      return [];
-    };
-
-    const selectedEquipment = getSportEquipment();
-
-    return (
-      <div className="p-8 text-center">
-        {/* Başarı ikonu */}
-        <div className="w-20 h-20 bg-green-100 rounded-full flex items-center justify-center mx-auto mb-6">
-          <svg className="w-10 h-10 text-green-500" fill="none" stroke="currentColor" viewBox="0 0 24 24">
-            <path strokeLinecap="round" strokeLinejoin="round" strokeWidth={2} d="M5 13l4 4L19 7" />
-          </svg>
-        </div>
-
-        {/* Başlık */}
-        <h3 className="text-2xl font-bold text-gray-800 mb-4">
-          Anketiniz Başarıyla Gönderildi! 🎉
-        </h3>
-
-        <p className="text-gray-600 mb-8">
-          Verdiğiniz cevaplar doğrultusunda size en uygun önerileri hazırlıyoruz. 
-          <br />
-          <strong>En kısa sürede size ulaşacağız!</strong>
-        </p>
-
-        {/* Anket özeti */}
-        <div className="bg-gray-50 rounded-xl p-6 mb-6 text-left">
-          <h4 className="font-semibold text-gray-800 mb-4 text-center">📋 Anket Özetiniz</h4>
-          
-          <div className="space-y-3">
-            {Boolean(submittedData?.['sport-type']) && (
-              <div className="flex justify-between">
-                <span className="text-gray-600">Spor Türü:</span>
-                <span className="font-medium text-gray-800">{String(submittedData?.['sport-type'])}</span>
-              </div>
-            )}
-
-            {selectedEquipment.length > 0 && (
-              <div>
-                <span className="text-gray-600 block mb-2">Seçilen Ekipmanlar:</span>
-                <div className="space-y-1">
-                  {selectedEquipment.map((equipment: string, index: number) => (
-                    <div key={index} className="text-sm text-gray-700 ml-4">
-                      • {equipment}
-                    </div>
-                  ))}
-                </div>
-              </div>
-            )}
-
-            {Boolean(submittedData?.intensity) && (
-              <div className="flex justify-between">
-                <span className="text-gray-600">Spor Yoğunluğu:</span>
-                <span className="font-medium text-gray-800">{String(submittedData?.intensity)}</span>
-              </div>
-            )}
-
-            {Boolean(submittedData?.['nutrition-goal']) && (
-              <div className="flex justify-between">
-                <span className="text-gray-600">Beslenme Hedefi:</span>
-                <span className="font-medium text-gray-800">{String(submittedData?.['nutrition-goal'])}</span>
-              </div>
-            )}
-
-            {/* Diğer anket türleri için dinamik içerik */}
-            {Boolean(submittedData?.['study-duration']) && (
-              <div className="flex justify-between">
-                <span className="text-gray-600">Çalışma Süresi:</span>
-                <span className="font-medium text-gray-800">{String(submittedData?.['study-duration'])}</span>
-              </div>
-            )}
-
-            {Boolean(submittedData?.['group-size']) && (
-              <div className="flex justify-between">
-                <span className="text-gray-600">Grup Büyüklüğü:</span>
-                <span className="font-medium text-gray-800">{String(submittedData?.['group-size'])}</span>
-              </div>
-            )}
-
-            {Boolean(submittedData?.['hunger-level']) && (
-              <div className="flex justify-between">
-                <span className="text-gray-600">Açlık Seviyesi:</span>
-                <span className="font-medium text-gray-800">{String(submittedData?.['hunger-level'])}</span>
-              </div>
-            )}
-          </div>
-        </div>
-
-        {/* İletişim bilgisi */}
-        <div className="bg-blue-50 rounded-xl p-6 mb-6">
-          <h4 className="font-semibold text-blue-800 mb-2">📞 İletişim</h4>
-          <p className="text-blue-700 text-sm">
-            Acil durumlar için WhatsApp: <strong>+90 5XX XXX XX XX</strong>
-            <br />
-            E-posta: <strong>info@sahilde.com</strong>
-          </p>
-        </div>
-
-        {/* Butonlar */}
-        <div className="flex gap-4 justify-center">
-          <Button
-            onClick={handleBackToSurvey}
-            variant="outline"
-            className="flex items-center space-x-2"
-          >
-            <ChevronLeft size={16} />
-            <span>Anketi Düzenle</span>
-          </Button>
-          
-          <Button
-            onClick={handleCloseModal}
-            className="bg-green-500 hover:bg-green-600 flex items-center space-x-2"
-          >
-            <span>Paketleri Görüntüle</span>
-            <ChevronRight size={16} />
-          </Button>
-        </div>
-      </div>
-    );
-  };
-
   if (!isOpen) return null;
 
   return (
@@ -372,7 +139,6 @@ export default function SurveyModal({ survey, isOpen, onClose, onSubmit }: Surve
                 <X size={24} />
               </button>
             </div>
-            
             {/* Progress Bar - sadece anket sırasında göster */}
             {!showSuccessPage && (
               <div className="mt-6">
@@ -394,61 +160,65 @@ export default function SurveyModal({ survey, isOpen, onClose, onSubmit }: Surve
 
           {/* Content */}
           {showSuccessPage ? (
-            renderSuccessPage()
+            <SurveySuccessPage
+              submittedData={submittedData}
+              handleBackToSurvey={handleBackToSurvey}
+              handleCloseModal={handleCloseModal}
+            />
           ) : (
-            <form onSubmit={handleSubmit(onFormSubmit)} className="p-6">
-              <AnimatePresence mode="wait">
-                <motion.div
-                  key={currentStep}
-                  initial={{ opacity: 0, x: 20 }}
-                  animate={{ opacity: 1, x: 0 }}
-                  exit={{ opacity: 0, x: -20 }}
-                  transition={{ duration: 0.3 }}
-                  className="space-y-6"
-                >
-                  <div>
-                    <h3 className="text-xl font-semibold text-gray-800 mb-4">
-                      {currentQuestion?.question}
-                      {currentQuestion?.required && <span className="text-red-500 ml-1">*</span>}
-                    </h3>
-                    {currentQuestion && renderQuestion(currentQuestion)}
-                  </div>
-                </motion.div>
-              </AnimatePresence>
-
-              {/* Navigation */}
-              <div className="flex justify-between items-center mt-8 pt-6 border-t border-gray-200">
-                <Button
-                  type="button"
-                  onClick={prevStep}
-                  disabled={isFirstStep}
-                  variant="outline"
-                  className="flex items-center space-x-2"
-                >
-                  <ChevronLeft size={16} />
-                  <span>Önceki</span>
-                </Button>
-
-                {isLastStep ? (
-                  <Button
-                    type="submit"
-                    className="flex items-center space-x-2 bg-green-500 hover:bg-green-600"
+            <FormProvider {...methods}>
+              <form onSubmit={handleSubmit(onFormSubmit)} className="p-6">
+                <AnimatePresence mode="wait">
+                  <motion.div
+                    key={currentStep}
+                    initial={{ opacity: 0, x: 20 }}
+                    animate={{ opacity: 1, x: 0 }}
+                    exit={{ opacity: 0, x: -20 }}
+                    transition={{ duration: 0.3 }}
+                    className="space-y-6"
                   >
-                    <Send size={16} />
-                    <span>Gönder</span>
-                  </Button>
-                ) : (
+                    <div>
+                      <h3 className="text-xl font-semibold text-gray-800 mb-4">
+                        {currentQuestion?.question}
+                        {currentQuestion?.required && <span className="text-red-500 ml-1">*</span>}
+                      </h3>
+                      {currentQuestion && <SurveyQuestionRenderer question={currentQuestion} />}
+                    </div>
+                  </motion.div>
+                </AnimatePresence>
+                {/* Navigation */}
+                <div className="flex justify-between items-center mt-8 pt-6 border-t border-gray-200">
                   <Button
                     type="button"
-                    onClick={nextStep}
+                    onClick={prevStep}
+                    disabled={isFirstStep}
+                    variant="outline"
                     className="flex items-center space-x-2"
                   >
-                    <span>Sonraki</span>
-                    <ChevronRight size={16} />
+                    <ChevronLeft size={16} />
+                    <span>Önceki</span>
                   </Button>
-                )}
-              </div>
-            </form>
+                  {isLastStep ? (
+                    <Button
+                      type="submit"
+                      className="flex items-center space-x-2 bg-green-500 hover:bg-green-600"
+                    >
+                      <Send size={16} />
+                      <span>Gönder</span>
+                    </Button>
+                  ) : (
+                    <Button
+                      type="button"
+                      onClick={nextStep}
+                      className="flex items-center space-x-2"
+                    >
+                      <span>Sonraki</span>
+                      <ChevronRight size={16} />
+                    </Button>
+                  )}
+                </div>
+              </form>
+            </FormProvider>
           )}
         </motion.div>
       </div>
